+++ pm.eclipse.editbox/src/mw/eclipse/TRC_Overlay/impl/BoxDecoratorImpl.java
@@ -1,4 +1,4 @@
-package mw.eclipse.editbox.impl;
+package mw.eclipse.TRC_Overlay.impl;
 
 import java.util.ArrayList;
 import java.util.Arrays;
