+++ pm.eclipse.editbox/src/mw/eclipse/TRC_Overlay/views/TRCView.java
@@ -50,7 +50,7 @@ import org.eclipse.ui.IWorkbenchWindow;
 import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.part.ViewPart;
 
-import mw.eclipse.editbox.impl.BoxDecoratorImpl;
+import mw.eclipse.TRC_Overlay.impl.BoxDecoratorImpl;
 import pm.eclipse.editbox.EditBox;
 import pm.eclipse.editbox.impl.TRCFileInteraction;
 import pm.eclipse.editbox.impl.TRCFileInteraction.TRCRequirement;
